move BulkEditFilter-related models to models/filters.py
